:sparkles: Support decimals in currency
This also adapts the prices of items to be more align with currencies like Euro or Dollars instead of having to make my own because of the lack of decimals.

+++ src/Entities/Types/Common.Types.fs
@@ -18,7 +18,7 @@ module CommonTypes =
     type times
 
     /// Defines an amount in the in-game currency.
-    type Amount = int<dd>
+    type Amount = decimal<dd>
 
     /// Measure for the quality of something, as a percentage.
     [<Measure>]
